♻️ useTaskLists에서 parseTasks를 유틸리티 함수로 분리
