+++ app/[teamId]/(index)/TaskProgressBadge.tsx
@@ -1,14 +1,13 @@
 import { ITask } from '@/types/task.type';
 import IconTodo from '@/public/images/icon-state-todo.svg';
 import IconDone from '@/public/images/icon-state-done.svg';
-import useTaskLists from '@/hooks/useTaskLists';
+import parseTasks from '@/utils/parseTasks';
 
 interface TaskProgressBadgeProps {
   tasks: ITask[];
 }
 
 export default function TaskProgressBadge({ tasks }: TaskProgressBadgeProps) {
-  const { parseTasks } = useTaskLists();
   const { length, done } = parseTasks(tasks);
 
   return (
